Main: make ResourceHandle size_t to avoid using 64bit atomics on 32bit

+++ OgreMain/include/OgreResource.h
@@ -35,7 +35,7 @@ THE SOFTWARE.
 
 namespace Ogre {
 
-    typedef unsigned long long int ResourceHandle;
+    typedef size_t ResourceHandle;
 
 
     // Forward declaration
